projects/gba-tester/requirements.txt
playwright
pytest-playwright
python-dotenv 